refactor(api-file-manager): consolidate into single FmGraphQLSchema
